26.08.2023

+++ src/main/java/com/example/cleverbankbyniunko/dao/AccountDao.java
@@ -8,6 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface AccountDao {
-    Optional<List<Account>> findAccountsByUserId(long userId) throws DaoException;
-    Optional<Account> selectAccountById(long id) throws DaoException;
+    Optional<List<Account>> findAccountsByUserId(int userId) throws DaoException;
+    Optional<Account> selectAccountById(int id) throws DaoException;
+
 }
